\name{d-methods}
\docType{methods}
\alias{d-methods}
\alias{d}
\alias{d,Distribution-method}
\title{ Methods for Function d in Package `distr' }
\description{d-methods}

\section{Methods}{\describe{

\item{d}{\code{signature(object = "Distribution")}: returns the density function }
}}
\keyword{}
