\name{df1-methods}
\docType{methods}
\alias{df1-methods}
\alias{df1<--methods}
\alias{df1}
\alias{df1<-}
\alias{df1,FParameter-method}
\alias{df1,F-method}
\alias{df1<-,FParameter-method}
\alias{df1<-,F-method}
\title{ Methods for Function df1 in Package `distr' }
\description{df-methods}

\section{Methods}{\describe{

\item{df1}{\code{signature(object = "FParameter")}: returns the slot \code{df1} of the parameter of an F-distribution }
\item{df1<-}{\code{signature(object = "FParameter")}: modifies the slot \code{df1} of the parameter of an F-distribution }

\item{df1}{\code{signature(object  = "F")}: returns the slot \code{df1} of the slot \code{param} of the distribution}
\item{df1<-}{\code{signature(object = "F")}: modifies the slot \code{df1} of the slot \code{param} of the distribution }
}}
\keyword{}
