\name{dimension-methods}
\docType{methods}
\alias{dimension-methods}
\alias{dimension<--methods}
\alias{dimension}
\alias{dimension<-}
\alias{dimension,EuclideanSpace-method}
\alias{dimension<-,EuclideanSpace-method}

\title{ Methods for Function dimension in Package `distr' }
\description{dimension-methods}
\section{Methods}{\describe{
    \item{dimension}{\code{signature(object = "EuclideanSpace")}: returns the dimension of the space }
    \item{dimension<-}{\code{signature(object = "EuclideanSpace")}: modifies the dimension of the space }
}}
\keyword{methods}
\concept{dimension}
\concept{space}
\concept{S4 space class}
\concept{accessor function}
\concept{replacement function}
