% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{distApply}
\alias{distApply}
\title{Computes the distributed R apply function}
\usage{
distApply(X, MARGIN, ...)
}
\arguments{
\item{X}{an array, including a matrix.}

\item{MARGIN}{a vector which specifies the direction to which the function
will be applied to X.}

\item{...}{Arguments of \link[base]{apply}}
}
\value{
token for the operation
}
\description{
This function works the same way as \link[base]{apply}
Returns an unique token for the operation, for more information
on getting the results of the operation, see \link{getResults}.
}
