% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{distLApply}
\alias{distLApply}
\title{Computes the distributed R lapply function.}
\usage{
distLApply(X, ...)
}
\arguments{
\item{X}{a vector (atomic or list) if not coerced to list}

\item{...}{Arguments of \link[base]{lapply}}
}
\value{
token for the operation
}
\description{
This function works the same way as \link[base]{lapply}
Returns an unique token for the operation, for more information
on getting the results of the operation, see \link{getResults}.
}
