% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{distR}
\alias{distR}
\title{distR}
\description{
Distributed system for computations. This package allows for
distributed computations in R's apply family of functions as well as
distributed computations writen by the user.This packages is to be used in
the client node when the worker nodes have been set up.
}
\section{Functions}{

\link{setupWorkspace}
\link{ExecuteLines}
\link{getResults}
\link{distApply}
\link{distTApply}
\link{distLApply}
\link{distSApply}
}

