\name{distrARITH}
\alias{distrARITH}
\alias{ARITHMETICS}
\title{Arithmetics on Distributions}
\description{Provides information on the interpretation of
arithmetics operating on Distributions in package \pkg{distr}}
\usage{
distrARITH()
}
\value{no value is returned
}

\author{Peter Ruckdeschel \email{Peter.Ruckdeschel@uni-bayreuth.de}}
\examples{
distrARITH()
}
\keyword{math}
\keyword{distribution}
\keyword{arith}

