\name{DistributionSymmetry-class}
\docType{class}
\alias{DistributionSymmetry-class}

\title{Class of Symmetries for Distributions}
\description{Class of symmetries for distributions.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{type}}{Object of class \code{"character"}:
      discribes type of symmetry. }
    \item{\code{SymmCenter}}{Object of class \code{"OptionalNumeric"}:
      center of symmetry. }
  }
}
\section{Extends}{
Class \code{"Symmetry"}, directly.
}
%\section{Methods}{}
%\references{}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
%\note{}
\seealso{\code{\link{Symmetry-class}}, \code{\link{Distribution-class}}, 
    \code{\link{OptionalNumeric-class}}}
%\examples{}
\concept{symmetry}
\keyword{classes}
\keyword{distribution}
