% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{distSApply}
\alias{distSApply}
\title{Computes the distributed R sapply function}
\usage{
distSApply(X, ...)
}
\arguments{
\item{X}{a vector (atomic or list) if not coerced to list}

\item{...}{Arguments of \link[base]{sapply}}
}
\value{
token for the operation
}
\description{
This function works the same way as \link[base]{sapply}
Returns an unique token for the operation, for more information
on getting the results of the operation, see \link{getResults}.
}
