\name{filename-methods}
\docType{methods}
\alias{filename-methods}
\alias{filename<--methods}
\alias{filename}
\alias{filename<-}
\alias{filename,Dataclass-method}
\alias{filename<-,Dataclass-method}
\alias{filename,Evaluation-method}

\title{ Methods for Function filename in Package `distr' }
\description{filename-methods}
\section{Methods}{\describe{

\item{filename}{\code{signature(object = "Dataclass")}: returns the filename }

\item{filename<-}{\code{signature(object = "Dataclass")}: changes the filename }
    
\item{filename}{\code{signature(object = "Evaluation")}: returns the filename of the evaluated object }
}}
\keyword{}
