% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getResults}
\alias{getResults}
\title{Obtain the results of a computation}
\usage{
getResults(token)
}
\arguments{
\item{token}{Token from a previous computation}
}
\value{
R object corresponding to the results
}
\description{
Given the token of a previous computation return the R object that results from the operation.
This function blocks the system until the data has been retrieved, therefore should be used when the data returned is necessary for the users script to continue.
}
