\name{k-methods}
\docType{methods}
\alias{k-methods}
\alias{k<--methods}
\alias{k}
\alias{k<-}
\alias{k,HyperParameter-method}
\alias{k,Hyper-method}
\alias{k<-,HyperParameter-method}
\alias{k<-,Hyper-method}
\title{ Methods for Function k in Package `distr' }
\description{k-methods}

\section{Methods}{\describe{

\item{k}{\code{signature(object = "HyperParameter")}: returns the slot \code{k} of the parameter of the distribution }
\item{k<-}{\code{signature(object = "HyperParameter")}: modifies the slot \code{k} of the parameter of the distribution }

\item{k}{\code{signature(object = "Hyper")}: returns the slot \code{k} of the parameter of the distribution }
\item{k<-}{\code{signature(object = "Hyper")}: modifies the slot \code{k} of the parameter of the distribution }
}}
\keyword{methods}
\keyword{distribution}
\concept{parameter}
\concept{Hypergeometric distribution}
\concept{S4 parameter class}
\concept{accessor function}
\concept{replacement function}
