\name{Length-methods}
\docType{methods}
\alias{Length-methods}
\alias{Length<--methods}
\alias{Length}
\alias{Length<-}
\alias{Length,Lattice-method}
\alias{Length<-,Lattice-method}
\alias{Length,LatticeDistribution-method}
\alias{Length<-,LatticeDistribution-method}
\title{ Methods for Function Length in Package `distr' }
\description{Length-methods}
\section{Methods}{\describe{
\item{Length}{\code{signature(object = "Lattice")}: returns the slot Length of the lattice }
\item{Length<-}{\code{signature(object = "Lattice")}: modifies the slot Length of the lattice }
\item{Length}{\code{signature(object = "LatticeDistribution")}: returns the slot Length of the lattice slot of the distribution}
\item{Length<-}{\code{signature(object = "LatticeDistribution")}: modifies the slot Length of the lattice slot of the distribution}
}}
\keyword{methods}
\keyword{distribution}
\concept{space}
\concept{S4 space class}
\concept{lattice}
\concept{affine linear lattice}
\concept{accessor function}
\concept{replacement function}
