\name{location-methods}
\docType{methods}
\alias{location-methods}
\alias{location<--methods}
\alias{location}
\alias{location<-}
\alias{location,LogisParameter-method}
\alias{location,Logis-method}
\alias{location,CauchyParameter-method}
\alias{location,Cauchy-method}
\alias{location,DiracParameter-method}
\alias{location,Dirac-method}
\alias{location<-,LogisParameter-method}
\alias{location<-,Logis-method}
\alias{location<-,CauchyParameter-method}
\alias{location<-,Cauchy-method}
\alias{location<-,DiracParameter-method}
\alias{location<-,Dirac-method}
\title{ Methods for Function location in Package `distr' }
\description{location-methods}

\section{Methods}{\describe{

\item{location}{\code{signature(object = "LogisParameter")}: returns the slot \code{location} of the parameter of the distribution }
\item{location<-}{\code{signature(object = "LogisParameter")}: modifies the slot \code{location} of the parameter of the
distribution }

\item{location}{\code{signature(object = "Logis")}: returns the slot \code{location} of the parameter of the distribution }
\item{location<-}{\code{signature(object = "Logis")}: modifies the slot \code{location} of the parameter of the distribution }

\item{location}{\code{signature(object = "CauchyParameter")}: returns the slot \code{location} of the parameter of the distribution }
\item{location<-}{\code{signature(object = "CauchyParameter")}: modifies the slot \code{location} of the parameter of the
distribution }

\item{location}{\code{signature(object = "Cauchy")}: returns the slot \code{location} of the parameter of the distribution }
\item{location<-}{\code{signature(object = "Cauchy")}: modifies the slot \code{location} of the parameter of the distribution }

\item{location}{\code{signature(object = "DiracParameter")}: returns the slot \code{location} of the parameter of the distribution }
\item{location<-}{\code{signature(object = "DiracParameter")}: modifies the slot \code{location} of the parameter of the
distribution}

\item{location}{\code{signature(object = "Dirac")}: returns the slot \code{location} of the parameter of the distribution }
\item{location<-}{\code{signature(object = "Dirac")}: modifies the slot \code{location} of the parameter of the distribution }
}}
\keyword{methods}
\keyword{distribution}
\concept{parameter}
\concept{Dirac distribution}
\concept{Cauchy distribution}
\concept{S4 parameter class}
\concept{accessor function}
\concept{replacement function}
\concept{location scale family}
\concept{location parameter}
