\name{Math-methods}
\docType{methods}
\alias{Math-methods}
\alias{Math,AbscontDistribution-method}
\alias{Math,DiscreteDistribution-method}
\title{ Methods for Function Math in Package `distr' }
\description{Math-methods}

\section{Methods}{\describe{
    \item{Math}{\code{signature(x = "AbscontDistribution")}: application of a
      mathematical function, e.g. sin or exp (does not work with log!), to this
      absolutely continouos distribution}
    \item{Math}{\code{signature(x = "AbscontDistribution")}: application of a
      mathematical function, e.g. sin or exp (does not work with log!), to this
      discrete distribution}
}}
\keyword{}
