\name{Max-methods}
\docType{methods}
\alias{Max-methods}
\alias{Max<--methods}
\alias{Max}
\alias{Max<-}
\alias{Max,UnifParameter-method}
\alias{Max,Unif-method}
\alias{Max<-,UnifParameter-method}
\alias{Max<-,Unif-method}
\title{ Methods for Function Max in Package `distr' }
\description{Max-methods}

\section{Methods}{\describe{

\item{Max}{\code{signature(object = "UnifParameter")}: returns the slot \code{Max} of the parameter of the distribution }
\item{Max<-}{\code{signature(object = "UnifParameter")}: modifies the slot \code{Max} of the parameter of the distribution }

\item{Max}{\code{signature(object = "Unif")}: returns the slot \code{Max} of the parameter of the distribution }
\item{Max<-}{\code{signature(object = "Unif")}: modifies the slot \code{Max} of the parameter of the distribution }
}}
\keyword{methods}
\keyword{distribution}
\concept{parameter}
\concept{Uniform distribution}
\concept{S4 parameter class}
\concept{accessor function}
\concept{replacement function}
