\name{meanlog-methods}
\docType{methods}
\alias{meanlog-methods}
\alias{meanlog<--methods}
\alias{meanlog}
\alias{meanlog<-}
\alias{meanlog,LnormParameter-method}
\alias{meanlog,Lnorm-method}
\alias{meanlog<-,LnormParameter-method}
\alias{meanlog<-,Lnorm-method}
\title{ Methods for Function meanlog in Package `distr' }
\description{meanlog-methods}
\section{Methods}{\describe{
\item{meanlog}{\code{signature(object = "LnormParameter")}: returns the slot meanlog of the parameter of the distribution }
\item{meanlog<-}{\code{signature(object = "LnormParameter")}: modifies the slot meanlog of the parameter of the distribution }
\item{meanlog}{\code{signature(object = "Lnorm")}: returns the slot meanlog of the parameter of the distribution }
\item{meanlog<-}{\code{signature(object = "Lnorm")}: modifies the slot meanlog of the parameter of the distribution }
}}
\keyword{methods}
\keyword{distribution}
\concept{parameter}
\concept{Log-Normal distribution}
\concept{S4 parameter class}
\concept{accessor function}
\concept{replacement function}
