\name{Min-methods}
\docType{methods}
\alias{Min-methods}
\alias{Min<--methods}
\alias{Min}
\alias{Min<-}
\alias{Min,UnifParameter-method}
\alias{Min,Unif-method}
\alias{Min<-,UnifParameter-method}
\alias{Min<-,Unif-method}
\title{ Methods for Function Min in Package `distr' }
\description{Min-methods}

\section{Methods}{\describe{

\item{Min}{\code{signature(object = "UnifParameter")}: returns the slot \code{Min} of the parameter of the distribution }
\item{Min<-}{\code{signature(object = "UnifParameter")}: modifies the slot \code{Min} of the parameter of the distribution }

\item{Min}{\code{signature(object = "Unif")}: returns the slot \code{Min} of the parameter of the distribution }
\item{Min<-}{\code{signature(object = "Unif")}: modifies the slot \code{Min} of the parameter of the distribution }
}}
\keyword{methods}
\keyword{distribution}
\concept{parameter}
\concept{Uniform distribution}
\concept{S4 parameter class}
\concept{accessor function}
\concept{replacement function}

