\name{name-methods}
\docType{methods}
\alias{name-methods}
\alias{name<--methods}
\alias{name}
\alias{name<-}
\alias{name,Parameter-method}
\alias{name<-,Parameter-method}
\alias{name,rSpace-method}
\alias{name<-,rSpace-method}
\alias{name,Evaluation-method}
\title{ Methods for Function name in Package `distr' }
\description{name-methods}
\section{Methods}{\describe{
\item{name}{\code{signature(object = "Parameter")}: returns the slot name of the parameter }
\item{name<-}{\code{signature(object = "Parameter")}: modifies the slot name of the parameter }
\item{name}{\code{signature(object = "rSpace")}: returns the slot name of the space }
\item{name<-}{\code{signature(object = "rSpace")}: modifies the slot name of the space }
\item{name}{\code{signature(object = "Evaluation")}: returns the slot name of data object }
}}
\keyword{}
