\name{ncp-methods}
\docType{methods}
\alias{ncp-methods}
\alias{ncp<--methods}
\alias{ncp}
\alias{ncp<-}
\alias{ncp,ChisqParameter-method}
\alias{ncp,Chisq-method}
\alias{ncp<-,ChisqParameter-method}
\alias{ncp<-,Chisq-method}
\title{ Methods for Function ncp in Package `distr' }
\description{ncp-methods}

\section{Methods}{\describe{

\item{ncp}{\code{signature(object = "ChisqParameter")}: returns the slot \code{ncp} of the parameter of the distribution }
\item{ncp<-}{\code{signature(object = "ChisqParameter")}: modifies the slot \code{ncp} of the parameter of the distribution }

\item{ncp}{\code{signature(object = "Chisq")}: returns the slot \code{ncp} of the parameter of the distribution }
\item{ncp<-}{\code{signature(object = "Chisq")}: modifies the slot \code{ncp} of the parameter of the distribution }
}}
\keyword{}
