\name{OptionalParameter-class}
\docType{class}
\alias{OptionalParameter-class}

\title{Class "OptionalParameter" }
\description{ auxiliary class; may contain either a Parameter or NULL, cf. J. Chambers, "green book".}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "OptionalParameter" in the signature.
}
\author{
  Thomas Stabla \email{statho3@web.de},\cr 
  Florian Camphausen \email{fcampi@gmx.de},\cr
  Peter Ruckdeschel \email{Peter.Ruckdeschel@uni-bayreuth.de},\cr
  Matthias Kohl \email{Matthias.Kohl@stamats.de}
}
\seealso{\code{\link{Parameter-class}}}
\keyword{}
