\name{p-methods}
\docType{methods}
\alias{p-methods}
\alias{p}
\alias{p,Distribution-method}
\title{ Methods for Function p in Package `distr' }
\description{p-methods}

\section{Methods}{\describe{

\item{p}{\code{signature(object = "Distribution")}: returns the cumulative distribution function }
}}
\keyword{}
