\name{p.l-methods}
\docType{methods}
\alias{p.l-methods}
\alias{p.l}
\alias{p.l,DiscreteDistribution-method}
\title{Methods for Function p.l in Package `distr' }
\description{p-methods}

\section{Methods}{\describe{
\item{p.l}{\code{signature(object = "DiscreteDistribution")}: 
   returns the left continuous cumulative distribution function,
   i.e.; \eqn{p.l(t) = P(object < t)}{p.l(t) = P(object < t)} }
}}
\seealso{
\code{\link{DiscreteDistribution-class}}
}
\keyword{distribution}
\keyword{methods}
\concept{cumulative distribution function}
\concept{CDF}
\concept{left-continuous}
