\name{param-methods}
\docType{methods}
\alias{param-methods}
\alias{param}
\alias{param,Distribution-method}
\title{ Methods for Function param in Package `distr' }
\description{param-methods}
\section{Methods}{\describe{
\item{param}{\code{signature(object = "Distribution")}: returns the parameter }
}}
\keyword{}
