\name{pivot-methods}
\docType{methods}
\alias{pivot-methods}
\alias{pivot<--methods}
\alias{pivot}
\alias{pivot<-}
\alias{pivot,Lattice-method}
\alias{pivot<-,Lattice-method}
\alias{pivot,LatticeDistribution-method}
\alias{pivot<-,LatticeDistribution-method}
\title{ Methods for Function pivot in Package `distr' }
\description{pivot-methods}
\section{Methods}{\describe{
\item{pivot}{\code{signature(object = "Lattice")}: returns the slot pivot of the lattice }
\item{pivot<-}{\code{signature(object = "Lattice")}: modifies the slot pivot of the lattice }
\item{pivot}{\code{signature(object = "LatticeDistribution")}: returns the slot pivot of the lattice slot of the distribution}
\item{pivot<-}{\code{signature(object = "LatticeDistribution")}: modifies the slot pivot of the lattice slot of the distribution}
}}
\keyword{methods}
\keyword{distribution}
\concept{space}
\concept{S4 space class}
\concept{lattice}
\concept{affine linear lattice}
\concept{accessor function}
\concept{replacement function}
