\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,AbscontDistribution-method}
\alias{plot,DiscreteDistribution-method}
\title{ Methods for Function plot in Package `distr' }
\description{plot-methods}
\section{Methods}{\describe{
\item{plot}{\code{signature(object = "AbscontDistribution")}: plots density, cumulative and the inverse of the cumulative function }
\item{plot}{\code{signature(object = "DiscreteDistribution")}: plots density, cumulative and the inverse of the cumulative function }
}}
\keyword{}
