\name{print-methods}
\docType{methods}
\alias{print-methods}
\alias{print,UnivariateDistribution-method}
\title{ Methods for Function print in Package `distr' }
\description{print-methods}

\section{Methods}{\describe{

\item{print}{\code{signature(x = "UnivariateDistribution")}: returns the class of the object and its parameters }

}}
\keyword{}
