\name{q-methods}
\docType{methods}
\alias{q-methods}
\alias{q}
\alias{q,Distribution-method}
\title{ Methods for Function q in Package `distr' }
\description{q-methods}

\section{Methods}{\describe{

\item{q}{\code{signature(object  = "Distribution")}: returns the quantile function }
}}
\keyword{}
