\name{r-methods}
\docType{methods}
\alias{r-methods}
\alias{r}
\alias{r,Distribution-method}
\title{ Methods for Function r in Package `distr' }
\description{r-methods}
\section{Methods}{\describe{
    \item{p}{\code{signature(object = "Distribution")}: returns the cumulative distribution function }
}}
\keyword{}
