\name{result-methods}
\docType{methods}
\alias{result-methods}
\alias{result}
\alias{result,Evaluation-method}
\title{ Methods for Function result in Package `distr'}
\description{result-methods}

\section{Methods}{\describe{
\item{result}{\code{signature(object = "Evaluation")}: returns the result of an evaluation}
}}
\keyword{}
