\name{scale-methods}
\docType{methods}
\alias{scale-methods}
\alias{scale<--methods}
\alias{scale}
\alias{scale<-}
\alias{scale,GammaParameter-method}
\alias{scale,Gamma-method}
\alias{scale,LogisParameter-method}
\alias{scale,Logis-method}
\alias{scale,WeibullParameter-method}
\alias{scale,Weibull-method}
\alias{scale,CauchyParameter-method}
\alias{scale,Cauchy-method}
\alias{scale<-,GammaParameter-method}
\alias{scale<-,Gamma-method}
\alias{scale<-,LogisParameter-method}
\alias{scale<-,Logis-method}
\alias{scale<-,WeibullParameter-method}
\alias{scale<-,Weibull-method}
\alias{scale<-,CauchyParameter-method}
\alias{scale<-,Cauchy-method}
\title{ Methods for Function scale in Package `distr' }
\description{scale-methods}

\section{Methods}{\describe{

\item{scale}{\code{signature(object = "GammaParameter")}: returns the slot \code{scale} of the parameter of the distribution }
\item{scale<-}{\code{signature(object = "GammaParameter")}: modifies the slot \code{scale} of the parameter of the distribution }

\item{scale}{\code{signature(object = "Gamma")}: returns the slot \code{scale} of the parameter of the distribution }
\item{scale<-}{\code{signature(object = "Gamma")}: modifies the slot \code{scale} of the parameter of the distribution }

\item{scale}{\code{signature(object = "LogisParameter")}: returns the slot \code{scale} of the parameter of the distribution }
\item{scale<-}{\code{signature(object = "LogisParameter")}: modifies the slot \code{scale} of the parameter of the distribution }

\item{scale}{\code{signature(object = "Logis")}: returns the slot \code{scale} of the parameter of the distribution }
\item{scale<-}{\code{signature(object = "Logis")}: modifies the slot \code{scale} of the parameter of the distribution }

\item{scale}{\code{signature(object = "WeibullParameter")}: returns the slot \code{scale} of the parameter of the distribution }
\item{scale<-}{\code{signature(object = "WeibullParameter")}: modifies the slot \code{scale} of the parameter of the distribution }

\item{scale}{\code{signature(object = "Weibull")}: returns the slot \code{scale} of the parameter of the distribution }
\item{scale<-}{\code{signature(object = "Weibull")}: modifies the slot \code{scale} of the parameter of the distribution }

\item{scale}{\code{signature(object = "CauchyParameter")}: returns the slot \code{scale} of the parameter of the distribution }
\item{scale<-}{\code{signature(object = "CauchyParameter")}: modifies the slot \code{scale} of the parameter of the distribution }

\item{scale}{\code{signature(object = "Cauchy")}: returns the slot \code{scale} of the parameter of the distribution }
\item{scale<-}{\code{signature(object = "Cauchy")}: modifies the slot \code{scale} of the parameter of the distribution }
}}
\keyword{}
