\name{sd-methods}
\docType{methods}
\alias{sd-methods}
\alias{sd<--methods}
\alias{sd}
\alias{sd<-}
\alias{sd,NormParameter-method}
\alias{sd,Norm-method}
\alias{sd<-,NormParameter-method}
\alias{sd<-,Norm-method}
\title{ Methods for Function sd in Package `distr' }
\description{sd-methods}
\section{Methods}{\describe{
\item{sd}{\code{signature(object = "NormParameter")}: returns the slot sd of the parameter of the distribution }    
\item{sd<-}{\code{signature(object = "NormParameter")}: modifies the slot sd of the parameter of the distribution }
\item{sd}{\code{signature(object = "Norm")}: returns the slot sd of the parameter of the distribution }
\item{sd<-}{\code{signature(object = "Norm")}: modifies the slot sd of the parameter of the distribution }
}}
\keyword{}
