% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setupWorkspace}
\alias{setupWorkspace}
\title{Sets up the configuration}
\usage{
setupWorkspace(filePath = "configuration.txt")
}
\arguments{
\item{filePath}{= "configuration.txt" Path to the input file}
}
\description{
Reads the configuration.txt file and configures the environment. For more
information read the setup documentation
}
