\name{shape-methods}
\docType{methods}
\alias{shape-methods}
\alias{shape<--methods}
\alias{shape}
\alias{shape<-}
\alias{shape,GammaParameter-method}
\alias{shape,Gamma-method}
\alias{shape,WeibullParameter-method}
\alias{shape,Weibull-method}
\alias{shape<-,GammaParameter-method}
\alias{shape<-,Gamma-method}
\alias{shape<-,WeibullParameter-method}
\alias{shape<-,Weibull-method}
\title{ Methods for Function shape in Package `distr' }
\description{shape-methods}

\section{Methods}{\describe{

\item{shape}{\code{signature(object  = "GammaParameter")}: returns the slot \code{shape} of a parameter of a Gamma distribution }
\item{shape<-}{\code{signature(object = "GammaParameter")}: modifies the slot \code{shape} of a parameter of a Gamma distribution }

\item{shape}{\code{signature(object = "Gamma")}: returns the slot \code{shape} of the parameter slot of a Gamma distribution }
\item{shape<-}{\code{signature(object = "Gamma")}: modifies the slot shape of the parameter slot of a Gamma distribution }

\item{shape}{\code{signature(object  = "WeibullParameter")}: returns the slot \code{shape} of a parameter of a Weibull distribution }
\item{shape<-}{\code{signature(object = "WeibullParameter")}: modifies the slot \code{shape} of a parameter of a Weibull distribution }

\item{shape}{\code{signature(object  = "Weibull")}: returns the slot \code{shape} of the parameter slot of the distribution }
\item{shape<-}{\code{signature(object = "Weibull")}: modifies the slot \code{shape} of the parameter slot of the distribution }
}}
\keyword{}
