\name{shape1-methods}
\docType{methods}
\alias{shape1-methods}
\alias{shape1<--methods}
\alias{shape1}
\alias{shape1<-}
\alias{shape1,BetaParameter-method}
\alias{shape1,Beta-method}
\alias{shape1<-,BetaParameter-method}
\alias{shape1<-,Beta-method}
\title{ Methods for Function shape1 in Package `distr' }
\description{shape-methods}

\section{Methods}{\describe{

\item{shape1}{\code{signature(object = "BetaParameter")}: returns the slot \code{shape1} of the parameter of the distribution }
\item{shape1<-}{\code{signature(object = "BetaParameter")}: modifies the slot \code{shape1} of the parameter of the distribution }

\item{shape1}{\code{signature(object = "Beta")}: returns the slot \code{shape1} of the parameter of the distribution }
\item{shape1<-}{\code{signature(object = "Beta")}: modifies the slot \code{shape1} of the parameter of the distribution }
}}
\keyword{distribution}
\keyword{methods}
\concept{shape}
\concept{parameter}
\concept{Beta distribution}
\concept{accessor function}
\concept{replacement function}
