\name{shape2-methods}
\docType{methods}
\alias{shape2-methods}
\alias{shape2<--methods}
\alias{shape2}
\alias{shape2<-}
\alias{shape2,BetaParameter-method}
\alias{shape2,Beta-method}
\alias{shape2<-,BetaParameter-method}
\alias{shape2<-,Beta-method}
\title{ Methods for Function shape2 in Package `distr' }
\description{shape-methods}
\section{Methods}{\describe{
\item{shape2}{\code{signature(object = "BetaParameter")}: returns the slot shape2 of the parameter of the distribution }
\item{shape2<-}{\code{signature(object = "BetaParameter")}: modifies the slot shape2 of the parameter of the distribution }
\item{shape2}{\code{signature(object = "Beta")}: returns the slot shape2 of the parameter of the distribution }
\item{shape2<-}{\code{signature(object = "Beta")}: modifies the slot shape2 of the parameter of the distribution }
}}
\keyword{distribution}
\keyword{methods}
\concept{shape}
\concept{parameter}
\concept{Beta distribution}
\concept{accessor function}
\concept{replacement function}
