\name{simulate-methods}
\docType{methods}
\alias{simulate-methods}
\alias{simulate}
\alias{simulate,Simulation-method}
\alias{simulate,Contsimulation-method}
\title{ Methods for Function simulate in Package `distr' }
\description{simulate-methods}

\section{Methods}{\describe{

\item{simulate}{\code{signature(x = "Simulation")}: generates the random numbers for the simulation }

\item{simulate}{\code{signature(x = "Contsimulation")}: generates the random numbers for the simulation }

}}
\keyword{}
