\name{support-methods}
\docType{methods}
\alias{support-methods}
\alias{support}
\alias{support,DiscreteDistribution-method}
\title{ Methods for Function support in Package `distr'}
\description{support-methods}

\section{Methods}{\describe{

\item{support}{\code{signature(object =
    "DiscreteDistributionContsimulation")}: returns the support}
}}
\keyword{}
