\name{UnivariateParameter-class}
\docType{class}
\alias{UnivariateParameter-class}
\alias{initialize,UnivariateParameter-method}

\title{Class "UnivariateParameter"}
\description{ \code{UnivariateParameter} is the mother-class of \code{AbscontParameter} and \code{DiscreteParameter}. }
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Slots}{
  \describe{
    \item{\code{name}:}{Object of class \code{"character"}: a name / comment for the parameters}
  }
}
\section{Extends}{
Class \code{"Parameter"}, directly.
}

\author{Thomas Stabla \email{Thomas.Stabla@uni-bayreuth.de},\cr Florian Camphausen \email{Florian.Camphausen@uni-bayreuth.de},\cr
 Peter Ruckdeschel \email{Peter.Ruckdeschel@uni-bayreuth.de},\cr Matthias Kohl \email{Matthias.Kohl@uni-bayreuth.de}}

\seealso{
\code{\link{Parameter-class}}
\code{\link{AbscontParameter-class}}
\code{\link{DiscreteParameter-class}}
\code{\link{UnivariateDistribution-class}}
}

\keyword{}
