\name{df2-methods}
\docType{methods}
\alias{df2-methods}
\alias{df2<--methods}
\alias{df2}
\alias{df2<-}
\alias{df2,FParameter-method}
\alias{df2,Fd-method}
\alias{df2<-,FParameter-method}
\alias{df2<-,Fd-method}
\title{ Methods for Function df2 in Package `distr' }
\description{df-methods}

\section{Methods}{\describe{

\item{df2}{\code{signature(object = "FParameter")}: returns the slot \code{df2} of the parameter of an F-distribution }
\item{df2<-}{\code{signature(object = "FParameter")}: modifies the slot \code{df2} of the parameter of an F-distribution }

\item{df2}{\code{signature(object  = "Fd")}: returns the slot \code{df2} of the slot \code{param} of the distribution}
\item{df2<-}{\code{signature(object = "Fd")}: modifies the slot \code{df2} of the slot \code{param} of the distribution }
}}
\keyword{distribution}
\keyword{methods}
\concept{degree of freedom}
\concept{parameter}
\concept{F distribution}
\concept{accessor function}
\concept{replacement function}
