\name{distrARITH}
\alias{distrARITH}
\alias{ARITHMETICS}
\title{Arithmetics on Distributions}
\description{Provides information on the interpretation of
arithmetics operating on Distributions in package \pkg{distr}}
\usage{
distrARITH(library = NULL)
}
\arguments{
  \item{library}{a character vector with path names of R libraries, or \code{NULL}.
                 The default value of \code{NULL} corresponds to all libraries
                 currently known.  If the default is used, the loaded packages
                 are searched before the libraries}
    }
\value{no value is returned
}

\author{Peter Ruckdeschel \email{peter.ruckdeschel@uni-oldenburg.de}}
\examples{
distrARITH()
}
\keyword{math}
\keyword{distribution}
\keyword{arith}
\keyword{documentation}
\concept{arithmetics for distributions}
\concept{info file}

