\name{p-methods}
\docType{methods}
\alias{p-methods}
\alias{p}
\alias{p,Distribution-method}
\title{ Methods for Function p in Package `distr' }
\description{p-methods}

\section{Methods}{\describe{

\item{p}{\code{signature(object = "Distribution")}: returns the cumulative distribution function,
i.e.; \eqn{p.l(t) = P(object \le t)}{p.l(t) = P(object <= t)}  }
}}
\seealso{
\code{\link{Distribution-class}}
}
\keyword{distribution}
\keyword{methods}
\concept{cumulative distribution function}
\concept{accessor function}
\concept{CDF}
\concept{right-continuous}
