\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,AbscontDistribution-method}
\alias{plot,DiscreteDistribution-method}
\alias{plot,Dataclass-method}
\alias{plot,Simulation-method}
\alias{plot,Contsimulation-method}
\alias{plot,Evaluation-method}
\title{ Methods for Function plot in Package `distr' }
\description{plot-methods}
\section{Methods}{\describe{
\item{plot}{\code{signature(object = "AbscontDistribution")}: plots density, cumulative and the inverse of the cumulative function }
\item{plot}{\code{signature(object = "DiscreteDistribution")}: plots density, cumulative and the inverse of the cumulative function }
\item{plot}{\code{signature(object = "Dataclass")}: produces a plot of the data matrix }
\item{plot}{\code{signature(object = "Simulation")}: produces a plot of the data matrix }
\item{plot}{\code{signature(object = "Contsimulation")}: produces a plot of the real data matrix }
\item{plot}{\code{signature(object = "Evaluation")}: returns a boxplot of the result }
}}
\keyword{}
