\name{r-methods}
\docType{methods}
\alias{r-methods}
\alias{r}
\alias{r,Distribution-method}
\title{ Methods for Function r in Package `distr' }
\description{r-methods}
\section{Methods}{\describe{
    \item{r}{\code{signature(object = "Distribution")}: generates random deviates
    according to the distribution}
}}
\seealso{
\code{\link{Distribution-class}}
}
\keyword{distribution}
\keyword{methods}
\concept{random number generator}
\concept{RNG}
\concept{accessor function}
\concept{pseudo random number}
\concept{anamorphosis}
